backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pymongo==4.5.0
python-dotenv==1.0.0
bson==0.5.10
flask-jwt-extended==4.5.2
pytest==7.4.0
black==23.7.0
